refactor: Add error type generic parameter to reader functions
This allows the reader functions to take any parser function that
returns a custom error type, the standard error type, or the anyhow
error type, improving flexibility and usability.

The only downside is that you may need to specify the error type
explicitly in some cases, but you can always use anyhow::Ok to avoid that.